+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - **IMPORTANT**: Switched from `MIT` to `LGPLv3` license
 - Switched from `instance_eval` to `define_singleton_method` to define attribute methods
 - Exclude unnecessary files from gem builds
+- Added retries count to result object instead of metadata
 
 ## [1.13.0] - 2025-12-23